requirements.txt
# AI & OCR Project Dependencies
# Install with: pip install -r requirements.txt

# API & HTTP
requests
fastapi
uvicorn[standard]
python-multipart

# Image Processing
opencv-python
Pillow
pyzbar

# Text-to-Speech
gtts
pyttsx3

# Configuration
python-dotenv